Move provenance definition to MultiMatrix
